+++ lib/github/plugin/test_client.ex
@@ -9,6 +9,12 @@ defmodule GitHub.Plugin.TestClient do
   alias GitHub.Plugin.TypedDecoder
   alias GitHub.Testing
 
+  @generic_documentation "https://docs.github.com/en/rest"
+  @not_found_message "Not Found"
+  @rate_limited_documentation "https://docs.github.com/rest/overview/resources-in-the-rest-api#rate-limiting"
+  @rate_limited_message "API rate limit exceeded for xxx.xxx.xxx.xxx. (But here's the good news: Authenticated requests get a higher rate limit. Check out the documentation for more details.)"
+  @unauthorized_message "Must have admin rights to [resource]"
+
   @doc """
   Perform a client operation using mocks, and record the call for later assertions
   """
@@ -35,12 +41,36 @@ defmodule GitHub.Plugin.TestClient do
 
   defp error(operation, :not_found, _opts) do
     error = %GitHub.BasicError{
-      documentation_url: "https://docs.github.com/rest/reference/repos#get-a-repository",
-      message: "Not Found"
+      documentation_url: @generic_documentation,
+      message: @not_found_message
     }
 
     TypedDecoder.normalize_errors(
       %Operation{operation | response_body: error, response_code: 404},
+      []
+    )
+  end
+
+  defp error(operation, :rate_limited, _opts) do
+    error = %GitHub.BasicError{
+      documentation_url: @rate_limited_documentation,
+      message: @rate_limited_message
+    }
+
+    TypedDecoder.normalize_errors(
+      %Operation{operation | response_body: error, response_code: 403},
+      []
+    )
+  end
+
+  defp error(operation, :unauthorized, _opts) do
+    error = %GitHub.BasicError{
+      documentation_url: @generic_documentation,
+      message: @unauthorized_message
+    }
+
+    TypedDecoder.normalize_errors(
+      %Operation{operation | response_body: error, response_code: 403},
       []
     )
   end